+++ dsmr/proto_client.go
@@ -4,32 +4,35 @@ import (
 	"context"
 	"fmt"
 
+	"google.golang.org/protobuf/proto"
+
 	"github.com/ava-labs/avalanchego/ids"
 	"github.com/ava-labs/avalanchego/network/p2p"
 	"github.com/ava-labs/avalanchego/utils/set"
+	"github.com/ava-labs/hypersdk/proto/pb/dsmr"
 )
 
-type Marshaler[T any, U any] interface {
+type marshaler[T any, U any] interface {
 	MarshalRequest(T) ([]byte, error)
 	UnmarshalResponse([]byte) (U, error)
 }
 
 type TypedClient[T any, U any] struct {
-	*p2p.Client
-	marshaler Marshaler[T, U]
+	client    *p2p.Client
+	marshaler marshaler[T, U]
 }
 
 // TODO merge upstream into avalanchego
-func NewTypedClient[T any, U any](client *p2p.Client, marshaler Marshaler[T, U]) *TypedClient[T, U] {
+func newTypedClient[T any, U any](client *p2p.Client, marshaler marshaler[T, U]) *TypedClient[T, U] {
 	return &TypedClient[T, U]{
-		Client:    client,
+		client:    client,
 		marshaler: marshaler,
 	}
 }
 
 // AppRequest issues an arbitrary request to a node.
 // [onResponse] is invoked upon an error or a response.
-func (c *TypedClient[T, U]) AppRequest(
+func (t *TypedClient[T, U]) AppRequest(
 	ctx context.Context,
 	nodeIDs set.Set[ids.NodeID],
 	request T,
@@ -40,7 +43,7 @@ func (c *TypedClient[T, U]) AppRequest(
 		err error,
 	)) error {
 	onByteResponse := func(ctx context.Context, nodeID ids.NodeID, responseBytes []byte, err error) {
-		response, parseErr := c.marshaler.UnmarshalResponse(responseBytes)
+		response, parseErr := t.marshaler.UnmarshalResponse(responseBytes)
 		if parseErr != nil {
 			//TODO how do we handle this?
 			return
@@ -50,15 +53,59 @@ func (c *TypedClient[T, U]) AppRequest(
 
 	}
 
-	requestBytes, err := c.marshaler.MarshalRequest(request)
+	requestBytes, err := t.marshaler.MarshalRequest(request)
 	if err != nil {
 		return fmt.Errorf("failed to marshal request: %w", err)
 	}
 
-	return c.Client.AppRequest(
+	return t.client.AppRequest(
 		ctx,
 		nodeIDs,
 		requestBytes,
 		onByteResponse,
 	)
+}
+
+type getChunkSignatureMarshaler struct{}
+
+func (g getChunkSignatureMarshaler) MarshalRequest(t *dsmr.GetChunkSignatureRequest) ([]byte, error) {
+	return proto.Marshal(t)
+}
+
+func (g getChunkSignatureMarshaler) UnmarshalResponse(bytes []byte) (*dsmr.GetChunkSignatureResponse, error) {
+	response := dsmr.GetChunkSignatureResponse{}
+	if err := proto.Unmarshal(bytes, &response); err != nil {
+		return nil, err
+	}
+
+	return &response, nil
+}
+
+type getChunkMarshaler struct{}
+
+func (g getChunkMarshaler) MarshalRequest(t *dsmr.GetChunkRequest) ([]byte, error) {
+	return proto.Marshal(t)
+}
+
+func (g getChunkMarshaler) UnmarshalResponse(bytes []byte) (*dsmr.GetChunkResponse, error) {
+	response := dsmr.GetChunkResponse{}
+	if err := proto.Unmarshal(bytes, &response); err != nil {
+		return nil, err
+	}
+
+	return &response, nil
+}
+
+func NewGetChunkClient(client *p2p.Client) *TypedClient[*dsmr.GetChunkRequest, *dsmr.GetChunkResponse] {
+	return &TypedClient[*dsmr.GetChunkRequest, *dsmr.GetChunkResponse]{
+		client:    client,
+		marshaler: getChunkMarshaler{},
+	}
+}
+
+func NewGetChunkSignatureClient(client *p2p.Client) *TypedClient[*dsmr.GetChunkSignatureRequest, *dsmr.GetChunkSignatureResponse] {
+	return &TypedClient[*dsmr.GetChunkSignatureRequest, *dsmr.GetChunkSignatureResponse]{
+		client:    client,
+		marshaler: getChunkSignatureMarshaler{},
+	}
 }
